+++ middleware/session.go
@@ -39,6 +39,7 @@ type sessRW struct {
 	r         *http.Request
 	domain    string
 	secretKey string
+	written   bool
 }
 
 var (
@@ -50,7 +51,7 @@ var (
 	_ http.Hijacker       = &sessRW{}
 	_ http.Pusher         = &sessRW{}
 	_ io.ReaderFrom       = &sessRW{}
-	// _ http.CloseNotifier  = &logRW{} // `http.CloseNotifier` has been deprecated sinc Go v1.11(year 2018)
+	// _ http.CloseNotifier  = &sessRW{} // `http.CloseNotifier` has been deprecated sinc Go v1.11(year 2018)
 )
 
 func newSessRW(
@@ -64,10 +65,11 @@ func newSessRW(
 		r:              r,
 		domain:         domain,
 		secretKey:      secretKey,
+		written:        false,
 	}
 }
 
-// Write recodes the size of bytes sent for logging purposes.
+// Write saves session data.
 func (srw *sessRW) Write(b []byte) (int, error) {
 	// 3. Save session cookie to response.
 
@@ -76,13 +78,16 @@ func (srw *sessRW) Write(b []byte) (int, error) {
 	// According to: https://pkg.go.dev/net/http#ResponseWriter
 	// Changing the header map after a call to WriteHeader/Write has no effect unless in some specific cases.
 	// Thus, we call sess.Save here just before any call to `ResponseWriter.Write` goes through.
-	sess.Save(
-		srw.r,
-		srw.ResponseWriter,
-		srw.domain,
-		sessionMaxAge,
-		srw.secretKey,
-	)
+	if !srw.written {
+		sess.Save(
+			srw.r,
+			srw.ResponseWriter,
+			srw.domain,
+			sessionMaxAge,
+			srw.secretKey,
+		)
+		srw.written = true
+	}
 
 	return srw.ResponseWriter.Write(b)
 }