+++ middleware/session_test.go
@@ -2,6 +2,7 @@ package middleware
 
 import (
 	"fmt"
+	"html/template"
 	"io"
 	"net/http"
 	"net/http/httptest"
@@ -28,6 +29,33 @@ func someSessionHandler(msg, key, value string) http.HandlerFunc {
 		sess.Set(r, key, value)
 		sess.SetM(r, bigMap())
 		fmt.Fprint(w, msg)
+	}
+}
+
+// See https://github.com/komuw/ong/issues/205
+func templateVarsHandler(t *testing.T, name string) http.HandlerFunc {
+	tmpl, err := template.New("myTpl").Parse(`<!DOCTYPE html>
+<html>
+<head>
+</head>
+<body>
+	<h2>Welcome to awesome website {{.Name}}.</h2>
+</body>
+</html>`)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	return func(w http.ResponseWriter, r *http.Request) {
+		sess.Set(r, "name", name)
+
+		data := struct {
+			Name string
+		}{Name: name}
+		if err = tmpl.Execute(w, data); err != nil {
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
 	}
 }
 
@@ -105,6 +133,32 @@ func TestSession(t *testing.T) {
 			attest.Subsequence(t, c.Value, key)
 			attest.Subsequence(t, c.Value, value)
 		}
+	})
+
+	t.Run("with template variables", func(t *testing.T) {
+		t.Parallel()
+
+		secretKey := "secretKey"
+		domain := "localhost"
+		name := "John Doe"
+		wrappedHandler := session(templateVarsHandler(t, name), secretKey, domain)
+
+		rec := httptest.NewRecorder()
+		req := httptest.NewRequest(http.MethodGet, "/someUri", nil)
+		wrappedHandler.ServeHTTP(rec, req)
+
+		res := rec.Result()
+		defer res.Body.Close()
+
+		rb, err := io.ReadAll(res.Body)
+		attest.Ok(t, err)
+
+		attest.Equal(t, res.StatusCode, http.StatusOK)
+		attest.Subsequence(t, string(rb), name)
+
+		attest.Equal(t, len(res.Cookies()), 1)
+		attest.Equal(t, res.Cookies()[0].Name, sess.CookieName)
+		attest.NotZero(t, res.Cookies()[0].Value)
 	})
 
 	t.Run("concurrency safe", func(t *testing.T) {